Backend/Hotel.BotWebApi/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
llama-cpp-python==0.2.90